source/vrf_fulfill_lambda/pyproject.toml
[tool.poetry]
name = "vrf_fulfill_lambda"
version = "0.1.0"
description = ""
authors = ["Eugene"]

[tool.poetry.dependencies]
python = "^3.9"
beartype = "^0.10.4"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
aws-lambda-powertools = "^1.28.0"  # CDK stack attaches `powertools` Lambda layer
boto3 = "^1.24.63"  # Lambda already has `boto3` installed

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
